Move TextFilterVisualsTest
